fix(general):Fixed use client pages, removed unnecessary packages and modules
