+++ package.json
@@ -9,7 +9,6 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@radix-ui/react-select": "^2.1.1",
     "@react-three/drei": "^9.105.6",
     "@react-three/fiber": "^8.17.8",
     "@tanstack/react-query": "^5.56.2",
