+++ src/app/[locale]/(auth)/layout.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import Image from "next/image";
 
 import mockifyBackground from "@/public/images/bg.jpg";
@@ -9,10 +7,8 @@ import "@/ui/styles/pages/auth.page.scss";
 
 export default function AuthLayout({
   children,
-  params: { locale },
 }: {
   children: React.ReactNode;
-  params: { locale: string };
 }) {
   return (
     <main className="auth-section">
