+++ src/app/[locale]/page.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import { useTranslations } from "next-intl";
 import { Link } from "@/navigation";
 import Image from "next/image";
